requirements.txt

# python
numpy
pandas
scipy
scikit-learn
matplotlib


einops
pytorch-lightning
tensorboard
torchmetrics

opencv-python
imageio
scikit-image
pywavelets

transformers
huggingface-hub


timm
omegaconf
segment-anything
pretrainedmodels
grad-cam

tqdm
requests
protobuf
six

hydra-core

sentencepiece


# conda install pytorch